ng: Some initial stubbing of the "components" configuration language
In many ways this is pretty similar to the configuration language for
inside a module, but has child component groups instead of child modules
and components instead of resources.

Maybe the exercise of building this out will help us see some more
commonality between these that can help us collapse these concepts
together more, but for now we're intentionally keeping them separate so
we can explore the specific new behaviors we're researching without
getting bogged down with the baggage of what the existing concepts
represent.